+++ CHANGELOG.md
@@ -9,6 +9,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 ### Added
 - `isBackgroundPlaybackEnabled` to `PlaybackConfig`. For now this is only supported on iOS.
 - `BackgroundPlayback` example to the example Application
+- Support for AirPlay on iOS
+  - `Player.isAirPlayActive` to indicate whether media is being played externally using AirPlay.
+  - `Player.isAirPlayAvailable` to indicate whether AirPlay is available.
+  - `Player.showAirPlayTargetPicker` to display the AirPlay playback target picker.
+  - `RemoteControlConfig.isAirPlayEnabled` to control whether AirPlay should be possible.
+  - `AirPlayAvailableEvent` which is emitted when AirPlay is available.
+  - `AirPlayChangedEvent` which is emitted when AirPlay playback starts or stops.
 
 ## [0.2.0] - 2023-11-06
 ### Added